backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
twilio==8.11.0
PyJWT==2.8.0
gunicorn==21.2.0